Create core.magics.auto according to new API.
