+++ IPython/core/interactiveshell.py
@@ -2005,7 +2005,7 @@ class InteractiveShell(SingletonConfigurable):
         self.register_magic_function = self.magics_manager.register_function
         self.define_magic = self.magics_manager.define_magic
 
-        self.register_magics(mf.AutoMagics, m.BasicMagics, m.CodeMagics,
+        self.register_magics(m.AutoMagics, m.BasicMagics, m.CodeMagics,
             m.ConfigMagics, mf.DeprecatedMagics, m.ExecutionMagics,
             mf.ExtensionsMagics, m.HistoryMagics, mf.LoggingMagics,
             m.NamespaceMagics, mf.OSMagics, mf.PylabMagics )
